classes/class-07-mcp-protocol/demos/langgraph-agent-with-mcp/requirements.txt
langchain-mcp-adapters
langgraph
langchain[openai]
python-dotenv>=1.0
streamlit>=1.32
